Start renaming the payunity connector to payone
This is just a start to rename all externally facing names from payunity
to payone, so the bundle can be installed in parallel with the existing
payunity one.

This is a fork of the payunity connector since the functionality will likely
be similar and the payunity one will be removed eventually, so this also
preserves the history.